+++ showcase/utils/__init__.py
@@ -7,15 +7,12 @@ from showcase.utils.expressions import (
     resolve_template,
 )
 from showcase.utils.langsmith import (
-    export_graph_png,
     get_client,
-    get_graph_mermaid,
     get_latest_run_id,
     get_project_name,
     get_run_url,
     is_tracing_enabled,
     log_execution,
-    print_graph_mermaid,
     print_run_tree,
 )
 from showcase.utils.logging import get_logger, setup_logging
@@ -36,10 +33,6 @@ __all__ = [
     "print_run_tree",
     "log_execution",
     "get_run_url",
-    # Graph visualization
-    "get_graph_mermaid",
-    "print_graph_mermaid",
-    "export_graph_png",
     # Logging
     "get_logger",
     "setup_logging",